타 로그인 유저 비디오 수정/삭제 라우팅 보호 추가

+++ controllers/videoController.js
@@ -69,7 +69,11 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator.toString() !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (e) {
     console.log(e);
     res.redirect(routes.home);
@@ -95,7 +99,12 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
-    await Video.findOneAndDelete({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator.toString() !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndDelete({ _id: id });
+    }
   } catch (e) {
     console.log(e);
   }
